Add CI support for MinGW-w64 and Intel ICX compilers (#1188)
This commit adds continuous integration testing for two additional
compiler toolchains:

- MinGW-w64: Tests with Python 3.12 on Windows using MSYS2
- Intel ICX: Tests with Python 3.12 on Ubuntu using the modern
  Clang-based Intel oneAPI compiler (with caching to reduce install time)

Additional changes:
- Update Python 3.14.0-rc.2 to 3.14 (official release)
- Remove disabled old-compilers job for Ubuntu 20.04
- Fix format string warning in scoped_pymalloc for MinGW (%zu -> %llu)
- Update documentation to list MinGW-w64 and Intel ICX as officially
  supported compilers

+++ src/nb_internals.h
@@ -493,7 +493,8 @@ template <typename T> struct scoped_pymalloc {
     scoped_pymalloc(size_t size = 1) {
         ptr = (T *) PyMem_Malloc(size * sizeof(T));
         if (!ptr)
-            fail("scoped_pymalloc(): could not allocate %zu bytes of memory!", size);
+            fail("scoped_pymalloc(): could not allocate %llu bytes of memory!",
+                 (unsigned long long) size);
     }
     ~scoped_pymalloc() { PyMem_Free(ptr); }
     T *release() {
